feat(remove): separate feature
